% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeprophage_matchsize_func.R
\name{activeprophage_matchsize_checker}
\alias{activeprophage_matchsize_checker}
\title{Active prophage shape-match size calculator}
\usage{
activeprophage_matchsize_checker(
  predictionsummary,
  prophageprediction_list,
  windowsize
)
}
\arguments{
\item{predictionsummary}{Prediction summary table}

\item{prophageprediction_list}{A list containing shape information associated with all contigs containing a potential active prophages. Generated with the allprophages_func}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Calculate the size, in base pairs, of the matching region for prophage shapes detected in whole-community read coverages
}
\keyword{internal}
