% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGapsAndElevs.R
\name{allGapsAndElevs}
\alias{allGapsAndElevs}
\title{Collects prophage classification shape-match information}
\usage{
allGapsAndElevs(bestMatchList)
}
\arguments{
\item{bestMatchList}{classifications made with shape_matcher function. classifications are stored as the first item in the bestMatchList.}
}
\description{
Collects shape information associated with all contigs predicted as containing prophages based on the results from the shape_matcher function.
}
\keyword{internal}
