% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateActiveProphageShapeMatches_func.R
\name{allprophages_func_WC}
\alias{allprophages_func_WC}
\title{Collects prophage prediction shape-match information}
\usage{
allprophages_func_WC(best_match_list)
}
\arguments{
\item{best_match_list}{Predictions made with shape_matcher function. Predictions are stored as the first item in the best_match_list.}
}
\description{
Collects shape information associated with all contigs predicted as containing prophages based on the results from the shape_matcher function.
}
