% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatInputFunctions.R
\name{changewindowSize}
\alias{changewindowSize}
\title{Change the pileup window size}
\usage{
changewindowSize(pileup, windowSize, mode)
}
\arguments{
\item{pileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{windowSize}{The number of basepairs to average read coverage values over.
Options are 100, 200, 500, 1000 ONLY. Default is 1000.}

\item{mode}{Either "genome" or "metagenome"}
}
\description{
Re-averages windows of pileup files with 100bp windows to reduce pileup size.
}
\keyword{internal}
