% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestMatchListFunctions.R
\name{classifQuality}
\alias{classifQuality}
\title{Determines pattern-match quality}
\usage{
classifQuality(bestMatchList, pileup, windowSize, mode)
}
\arguments{
\item{bestMatchList}{A list containing pattern-match information associated with
all contigs/chunks classified by `ProActive()` pattern-matching}

\item{pileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{windowSize}{The number of basepairs to average read coverage values over.
Options are 100, 200, 500, 1000 ONLY. Default is 1000.}

\item{mode}{Either "genome" or "metagenome".}
}
\description{
The quality is either 'VeryHigh', 'High', or 'Low'. Pattern-matches
receive a very high quality rating when all the lowest or highest coverage values on
the contig/chunk fall with the gapped or elevated region, respectively, of the
pattern-match. Pattern-matches receive a high quality rating when the contig/chunk's
maximum or minimum read coverage value falls within the elevated or gapped region,
respectively, of the pattern-match. Pattern-matches receive a low quality rating
when the contig/chunk's maximum or minimum read coverage value does not fall within
the elevated or gapped region, respectively, of the pattern-match.
}
\keyword{internal}
