% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizesActiveProphagePredictions_func.R
\name{contig_prediction_summary_WC}
\alias{contig_prediction_summary_WC}
\title{Summarizes best shape-matches into a table}
\usage{
contig_prediction_summary_WC(best_match_list)
}
\arguments{
\item{best_match_list}{Predictions made with shape_matcher function. Predictions are stored as the first item in the best_match_list.}
}
\description{
Summarizes the predictions made in the shape_matcher function based on which shape match achieved the lowest mean absolute difference for each contig. Outputs results in a table rather than a list.
}
