% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialElevGapFunctions.R
\name{elevOrGapClassif}
\alias{elevOrGapClassif}
\title{Classifies partial elevation/gap pattern-matches}
\usage{
elevOrGapClassif(bestMatchList, pileupSubset)
}
\arguments{
\item{bestMatchList}{A list containing pattern-match information associated with
all contigs/chunks classified by `ProActive()` pattern-matching}

\item{pileupSubset}{A subset of the pileup that pertains only to the contig/chunk
currently being assessed.}
}
\description{
classify the contig/chunk as 'gap' if the elevated region is less than 50% of
the length of the contig/chunk and otherwise classify as 'elevation'.
}
\keyword{internal}
