% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleGenomegffTSVSubset.R
\docType{data}
\name{exampleGenomegffTSVSubset}
\alias{exampleGenomegffTSVSubset}
\title{exampleGenomegffTSVSubset}
\format{
## `exampleGenomegffTSVSubset`
A data frame with 306 rows and 9 columns:
\describe{
  \item{V1}{seqid}
  \item{V2}{source}
  \item{V3}{type}
  \item{V4}{start}
  \item{V5}{end}
  \item{V6}{score}
  \item{V7}{strand}
  \item{V8}{phase}
  \item{V9}{attributes}
}
}
\usage{
exampleGenomegffTSVSubset
}
\description{
Gene annotation subset associated with the genome in the sampleGenomePileup
Report...
}
\details{
This dataset is a subset of the sampleGenomePileupgffTSV generated by running:
"exampleGenomegffTSVSubset <- sampleGenomegffTSV[which(sampleGenomegffTSV[,5]<=350000),]"
}
\keyword{internal}
