% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleGenomePileupSubset.R
\docType{data}
\name{exampleGenomePileupSubset}
\alias{exampleGenomePileupSubset}
\title{exampleGenomePileupSubset}
\format{
## `exampleGenomePileupSubset`
A data frame with 3,500 rows and 4 columns:
\describe{
  \item{V1}{Accession}
  \item{V2}{Mapped read coverage averaged over a 100 bp window size}
  \item{V3}{Starting position (bp) of each 100 bp window. Starts from 100.}
  \item{V4}{Starting position (bp) of each 100 bp window. Starts from 0.}
}
}
\usage{
exampleGenomePileupSubset
}
\description{
A subset of the sampleGenomePileup used to demonstrate code in examples in
under 10 seconds.
Report...
}
\details{
This dataset is a subset of the sampleGenomePileup generated by running:
"exampleGenomePileupSubset <- sampleGenomePileup[which(sampleGenomePileup[,3]<=350000),]"
}
\keyword{internal}
