% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplePlotsGenomeSubset.R
\docType{data}
\name{examplePlotsGenomeSubset}
\alias{examplePlotsGenomeSubset}
\title{examplePlotsGenomeSubset}
\format{
## `examplePlotsGenomeSubset`
A list with 2 ggplot objects.
}
\usage{
examplePlotsGenomeSubset
}
\description{
Gene annotations associated with the metagenome in the exampleGenomePileupSubset
Report...
}
\details{
Output from running `plotProActiveResults()` on the exampleGenomePileupSubset
and the output of `ProActive()` run with the exampleGenomePileupSubset. Default
parameters used.
}
\keyword{internal}
