% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullElevGapShrink.R
\name{fullElevGapShrink}
\alias{fullElevGapShrink}
\title{Build elevation pattern, translate it across contig and shrink the width}
\usage{
fullElevGapShrink(
  minCov,
  windowSize,
  maxCov,
  elevLength,
  nonElev,
  bestMatchInfo,
  pileupSubset,
  minSize,
  elevOrGap
)
}
\arguments{
\item{pileupSubset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}
}
\description{
Build full-size elevation pattern, run it through the elevation translator function and then shrink the pattern until it reaches the minSize
}
\keyword{internal}
