% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeChunks.R
\name{genomeChunks}
\alias{genomeChunks}
\title{Break genome pileup into chunks for pattern-matching}
\usage{
genomeChunks(genomePileup, chunkSize)
}
\arguments{
\item{genomePileup}{Pileup file for read coverages mapped to a genome}

\item{chunkSize}{Value to chunk genome into. Should be divisible by 100. Default is 100,000bp per chunk.}
}
\description{
Break genome pileup into chunks for pattern-matching
}
\keyword{internal}
