% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPattern.R
\name{noPattern}
\alias{noPattern}
\title{No read coverage pattern}
\usage{
noPattern(pileupSubset)
}
\arguments{
\item{pileupSubset}{A subset of the read coverage dataset that pertains only to
the contig currently being assessed}
}
\description{
Assess whether a contig/chunk does not have an elevated/gapped read coverage pattern.
A horizontal line at the mean or median coverage should be the best
match if the contig/chunk read coverage is not gapped or elevated.
}
\keyword{internal}
