% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noactiveprophage_func.R
\name{notransduction_func_WC}
\alias{notransduction_func_WC}
\title{No transduction shape-matcher for whole-community read coverage}
\usage{
notransduction_func_WC(microbial_subset)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}
}
\description{
Assess whether a contig does not have a read coverage pattern associated with a transduction event. A horizontal line at the mean coverage should be an optimal match if the contig read coverage displays no transduction patterns
}
