% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialElevGap.R
\name{partialElevGap}
\alias{partialElevGap}
\title{Prophage off-left pattern-matcher for whole-community read coverage}
\usage{
partialElevGap(pileupSubset, windowSize, minSize, maxSize, elevOrGap)
}
\arguments{
\item{pileupSubset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowSize}{The window size used to re-average read coverage datasets}

\item{minSize}{The minimum size of elevated read coverage that ProActive searches for. Default is 10000 base pairs.}

\item{maxSize}{The minimum size of elevated read coverage that ProActive searches for. Default is the close to the length of the contig being assessed.}
}
\description{
Build and translate and prophage pattern going off the left side of the graph across a contig. Stop translating when the pattern left on the contig is 10,000bp. Translate the pattern 2000bp at a time.
Elevation pattern off left or right side
}
\keyword{internal}
