% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBuilder.R
\name{patternBuilder}
\alias{patternBuilder}
\title{Builds pattern-match vectors}
\usage{
patternBuilder(pileupSubset, bestMatchList)
}
\arguments{
\item{pileupSubset}{A subset of the pileup that pertains only to the contig/chunk
currently being assessed.}

\item{bestMatchList}{A list containing pattern-match information associated with
all contigs/chunks classified by `ProActive()` pattern-matching}
}
\description{
Builds the pattern-match (vector) associated with each contig/chunk for
visualization.
}
\keyword{internal}
