% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternTranslator.R
\name{patternTranslator}
\alias{patternTranslator}
\title{Full elevation pattern pattern-translator for whole-community read coverage only}
\usage{
patternTranslator(contigCov, bestMatchInfo, windowSize, pattern, elevOrGap)
}
\arguments{
\item{contigCov}{The read coverages that pertains only to the contig currently being assessed}

\item{bestMatchInfo}{The information associated with the current best pattern match. Includes the comparison score, the minimum and maximum pattern y-acis values, the pattern length, and the pattern start and stop positions on the contig}

\item{windowSize}{The window size used to re-average read coverage datasets}

\item{pattern}{A vector containing the values associated with the elevation pattern to be translated across the contig}

\item{elevOrGap}{}
}
\description{
Translates full elevation pattern across a contig. Translate the pattern 1000bp at a time. Stop translating when the elevation pattern is 5000bp from the end of the contig.
}
\keyword{internal}
