% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateProphageGeneAnnots_func.R
\name{phage_gene_finder}
\alias{phage_gene_finder}
\title{Locate phage-related gene annotations on contigs containing potential active prophages}
\usage{
phage_gene_finder(
  cleaned_gene_annots,
  prophagepredictions,
  microbialread_dataset
)
}
\arguments{
\item{cleaned_gene_annots}{A table of gene annotations made by PROKKA that have been passed through the prokka_cleanup function.}

\item{prophagepredictions}{A list containing shape information associated with all contigs containing a potential active prophage. The third item in the list output from the main prophage_activity_finder_func}

\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}
}
\description{
Search contigs containing potential active prophages (detected in whole-community read coverage only) for gene-annotations containing the words: "phage","bacteriophage","portal","spike", "tail", "capsid", "holin", "integrase", "needle", and "lysin". The matching gene-annotation locations are plotted on the associated read coverage graph
}
\examples{
\dontrun{
phage_gene_finder(gene_annots_cleaned, active_prophage_results[[3]], whole_commreadcoverages)
}
}
