% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatInputFunctions.R
\name{pileupFormatter}
\alias{pileupFormatter}
\title{Reformat input pileup file}
\usage{
pileupFormatter(pileup, mode)
}
\arguments{
\item{pileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{mode}{Either "genome" or "metagenome"}
}
\description{
Place columns in correct order, clean accessions by removing text after white
space, and name columns
}
\keyword{internal}
