% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProActive.R
\name{ProActive}
\alias{ProActive}
\title{Detect elevations and gaps in mapped read coverage patterns.}
\usage{
ProActive(
  pileup,
  mode,
  gffTSV,
  windowSize = 1000,
  minSize = 10000,
  maxSize = Inf,
  minContigLength = 30000,
  chunkSize = 50000,
  IncludeNoPatterns = FALSE,
  saveFilesTo
)
}
\arguments{
\item{pileup}{A pileup file for either a genome or metagenome containing mapped read coverages averaged over 100 bp window sizes}

\item{mode}{Either "genome" or "metagenome"}

\item{gffTSV}{Optional,}

\item{windowSize}{The number of base pairs to average coverage values over. Must be either 200, 500, or 1000. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. 1000 is the default.}

\item{minSize}{The minimum size (in bp) of elevated/gapped read coverage to detect. Default is 10000.}

\item{maxSize}{The minimum size (in bp) of elevated/gapped read coverage to detect. Default is NA (i.e no maximum).}

\item{chunkSize}{If mode="genome", ProActive will split your genome into chunks to search for elevated read coverage. Since ProActive can only identify one elevated read coverage pattern per 'chunk', choosing smaller chunks will increase sensitivity AND processing time. Default is 100000 (bp per chunk).}

\item{IncludeNoPatterns}{TRUE or FALSE. Default is FALSE. TRUE if you would like to include the contigs classified as 'none' in your final pattern lists. This can be useful if you are comparing ProActive results between conditions.}

\item{saveFilesTo}{Optional, Provide a path to the directory you wish to save output to. A folder will be made within the provided directory to store results.}
}
\description{
Performs read coverage pattern-matching and summarizes the results into a list.
The first item in the list is a table consisting of the summary information of all the
contigs that passed through shape-matching (i.e were not filtered out). The second item
in the list is a table consisting of the summary information of all contigs that were
predicted as containing a potential active prophage. The third item in the list contains
the best shape-match information associated with each contig in the previous table. The
fourth and final object in the list is a table containing the contigs that were filtered
out prior to shape_matching and the reason why.
}
\examples{
\dontrun{
ProActive_results <- ProActive(pileup=whole_commreadcoverages, mode="metagenome")
}
}
