% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProkkaGeneAnnotCleanup_func.R
\name{prokka_cleanup}
\alias{prokka_cleanup}
\title{Prokka gff file reformatter}
\usage{
prokka_cleanup(gff_file)
}
\arguments{
\item{gff_file}{A gff file containing gene annotation information from Prokka. Your gff file should have columns in the following order: contig (i.e "NODE_#"), tool (i.e "prodigal"), CDS, start position, stop position, '.', '+/-', '0', annotations.}
}
\description{
Prokka gene annotations cleanup
}
