% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeprophage_offrightside_func.R
\name{prophage_off_right_func_WC}
\alias{prophage_off_right_func_WC}
\title{Prophage off-right shape-matcher}
\usage{
prophage_off_right_func_WC(microbial_subset, windowsize)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Build and translate and prophage shape going off the right side of the graph across a contig. Stop translating when the shape left on the contig is 10,000bp. Translate the shape 2000bp at a time.
}
