% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helppage.R
\docType{package}
\name{ProphageActivityTool}
\alias{ProphageActivityTool}
\title{ProActive}
\description{
Detects active prophages in mapped read coverages from metagenomic sequencing using a reference-independent shape-matching approach.
}
\author{
Jessie Maier
}
