% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophage_shape_translator_func_activeprophage.R
\name{proshape_translator_func_WC}
\alias{proshape_translator_func_WC}
\title{Full prophage shape shape-translator for whole-community read coverage only}
\usage{
proshape_translator_func_WC(
  microbial_subset,
  best_match_info,
  windowsize,
  shape
)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{best_match_info}{The information associated with the current best shape match. Includes the comparison score, the minimum and maximum shape y-acis values, the shape length, and the shape start and stop positions on the contig}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{shape}{A vector containing the values associated with the prophage shape to be translated across the contig}
}
\description{
Translates full prophage shape across a contig. Translate the shape 2000bp at a time. Stop translating when the prophage shape is 5000bp from the end of the contig.
}
