% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCgap_off_right_func.R
\name{RCgap_off_right_func}
\alias{RCgap_off_right_func}
\title{Read coverage gap off-right shape-matcher}
\usage{
RCgap_off_right_func(microbial_subset, windowsize, minsize, maxsize)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{minsize}{The minimum size of elevated read coverage that ProActive searches for. Default is 10000 base pairs.}

\item{maxsize}{The minimum size of elevated read coverage that ProActive searches for. Default is the close to the length of the contig being assessed.}
}
\description{
Build and translate read coverage gap shape going off the right side of the graph across a contig. Stop translating when the shape left on the contig is 10,000bp. Translate the shape 2000bp at a time.
}
\keyword{internal}
