% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCgap_translator_func.R
\name{RCgap_translator_func}
\alias{RCgap_translator_func}
\title{Full read coverage gap shape-translator for whole-community read coverage only}
\usage{
RCgap_translator_func(Cov_values_contig, best_match_info, windowsize, shape)
}
\arguments{
\item{Cov_values_contig}{The read coverages that pertains only to the contig currently being assessed}

\item{best_match_info}{The information associated with the current best shape match. Includes the comparison score, the minimum and maximum shape y-axis values, the shape length, and the shape start and stop positions on the contig}

\item{windowsize}{The window size used to re-average read coverage datasets}

\item{shape}{A vector containing the values associated with the prophage shape to be translated across the contig}
}
\description{
Translates full gap shape across a contig. Translate the shape 2000bp at a time. Stop translating when the gap shape is 5000bp from the end of the contig.
}
\keyword{internal}
