% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestMatchListFunctions.R
\name{removeNoPatterns}
\alias{removeNoPatterns}
\title{Removes 'NoPattern' classifications from best match list}
\usage{
removeNoPatterns(bestMatchList)
}
\arguments{
\item{bestMatchList}{A list containing pattern-match information associated with
all contigs/chunks classified by `ProActive()` pattern-matching}
}
\description{
Removes 'NoPattern' classifications from the list of pattern-match information
associated with the best pattern-matches for each contig/chunk
}
\keyword{internal}
