% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGenomePileup.R
\docType{data}
\name{sampleGenomePileup}
\alias{sampleGenomePileup}
\title{sampleGenomePileup}
\format{
## `sampleGenomePileup`
A data frame with 48,575 rows and 4 columns:
\describe{
  \item{V1}{Accession}
  \item{V2}{Mapped read coverage averaged over a 100 bp window size}
  \item{V3}{Starting position (bp) of each 100 bp window. Starts from 100.}
  \item{V4}{Starting position (bp) of each 100 bp window. Starts from 0.}
}
}
\source{
<https://pubmed.ncbi.nlm.nih.gov/25608871/>
}
\usage{
sampleGenomePileup
}
\description{
A pileup file generated during read mapping to the *Salmonella enterica* LT2 genome.
Report...
}
\details{
This dataset was generated by extracting DNA from a culture of
*Salmonella enterica* LT2 (LT2) infected with phage P22. The DNA was shotgun
sequenced with Illumina (paired-end mode, 150 bp reads). The sequencing reads
were mapped to the LT2 reference genome (NCBI RefSeq NC_003197.2). The
bbmap.sh bincov parameter with covbinsize=100 was used to create a pileup file
with 100 bp windows.
}
\keyword{internal}
