% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMetagenomegffTSV.R
\docType{data}
\name{sampleMetagenomegffTSV}
\alias{sampleMetagenomegffTSV}
\title{sampleMetagenomegffTSV}
\format{
## `sampleMetagenomegffTSV`
A data frame with 467 rows and 9 columns:
\describe{
  \item{V1}{seqid}
  \item{V2}{source}
  \item{V3}{type}
  \item{V4}{start}
  \item{V5}{end}
  \item{V6}{score}
  \item{V7}{strand}
  \item{V8}{phase}
  \item{V9}{attributes}
}
}
\source{
<https://microbiomejournal.biomedcentral.com/articles/10.1186/s40168-020-00935-5>
}
\usage{
sampleMetagenomegffTSV
}
\description{
A subset of gene annotations associated with the metagenome in the sampleMetagenomePileup
Report...
}
\details{
This is a standard .gff file format. The .gff file was generated by
running PROKKA with default parameters on the metagenome assembly
associated with the sampleMetagenomePileup in the ProActive package.
The gff was subset to only include the data associated with the
contigs in the sample data subset.
}
\keyword{datasets}
