% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMetagenomeResults.R
\docType{data}
\name{sampleMetagenomeResults}
\alias{sampleMetagenomeResults}
\title{sampleMetagenomeResults}
\format{
## `sampleMetagenomeResults`
A list with 6 objects:
\describe{
  \item{SummaryTable}{A table containing all pattern-matching classifications}
  \item{CleanSummaryTable}{A table containing only gap and elevation pattern-match
classifications (i.e. noPattern classifications removed)}
  \item{PatternMatches}{A list object containing information needed to visualize the
pattern-matches in `plotProActiveResults()`}
  \item{FilteredOut}{A table containing contigs/chunks that were filtered out for being
too small or having too low read coverage}
  \item{Arguments}{A list object containing arguments used for pattern-matching (windowSize,
mode, chunkSize, chunkContigs)}
  \item{GeneAnnotTable}{A table containing gene predictions associated with elevated
or gapped regions in pattern-matches}
}
}
\usage{
sampleMetagenomeResults
}
\description{
Output of `ProActiveDetect()`
Report...
}
\details{
This data was generated by running `ProActiveDetect()` on the sampleMetagenomePileup
and sampleMetagenomegffTSV with default parameters.
}
\keyword{internal}
