% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePlotsGenome.R
\docType{data}
\name{samplePlotsGenome}
\alias{samplePlotsGenome}
\title{samplePlotsGenome}
\format{
## `samplePlotsGenome`
A list with 28 ggplot objects.
}
\usage{
samplePlotsGenome
}
\description{
Gene annotations associated with the metagenome in the sampleGenomePileup
Report...
}
\details{
Output from running `plotProActiveResults()` on the sampleGenomePileup
and the output of `ProActive()` run with the sampleGenomePileup. Default
parameters used.
}
\keyword{datasets}
