% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePlotsMetagenome.R
\docType{data}
\name{samplePlotsMetagenome}
\alias{samplePlotsMetagenome}
\title{samplePlotsMetagenome}
\format{
## `samplePlotsMetagenome`
A list with 6 ggplot objects.
}
\usage{
samplePlotsMetagenome
}
\description{
Gene annotations associated with the metagenome in the sampleGenomePileup
Report...
}
\details{
Output from running `plotProActiveResults()` on the sampleMetagenomePileup
and the output of `ProActive()` run with the sampleMetagenomePileup. Default
parameters used.
}
\keyword{internal}
