% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_builder_func.R
\name{shape_builder_func}
\alias{shape_builder_func}
\title{Shape-builder for patterns detected in whole-community read coverages}
\usage{
shape_builder_func(microbial_subset, ProActivePredictions, i)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{ProActivePredictions}{A list containing shape information associated with all contigs classified by ProActive}

\item{i}{The index position associated with the current contig's best shape-match information}
}
\description{
Builds the shape (vector) associated with the 'best shape-match' information associated with each contig so it can be plotted.
}
\keyword{internal}
