% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildActiveProphageShapes_func.R
\name{shape_builder_func_WC}
\alias{shape_builder_func_WC}
\title{Shape-builder for active prophages detected in whole-community read coverages}
\usage{
shape_builder_func_WC(microbial_subset, prophagepredictions, i)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{prophagepredictions}{A list containing shape information associated with all contigs containing a potential active prophage. Generated with the allprophages_func}

\item{i}{The index position associated with the current contig's best shape-match information}
}
\description{
Builds the shape (vector) associated with the 'best shape-match' information associated with each contig so it can be plotted.
}
