% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeWindowSize.R
\name{changewindowSize}
\alias{changewindowSize}
\title{Change the read coverage window size}
\usage{
changewindowSize(pileup, windowSize, mode)
}
\arguments{
\item{pileup}{A read coverage dataset that has been cleaned and reformatted by the readcovdf_formatter function}

\item{windowSize}{The number of base pairs to average coverage values over}

\item{mode}{Either "genome" or "metagenome"}
}
\description{
Re-averages window sizes of read coverage averages. Start with 100bp windows always. Cannot make window size less than 100bp.
}
\keyword{internal}
