% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifQuality.R
\name{classifQuality}
\alias{classifQuality}
\title{Checks for max or minimum coverage outside match region}
\usage{
classifQuality(bestMatchList, pileup, windowSize, mode)
}
\arguments{
\item{windowSize}{The number of base pairs to average coverage values over. For compatibility with shape-matching, window sizes can only be 200, 500, 1000 and 2000. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. 1000bp windowSize is the default.}

\item{mode}{Either "genome" or "metagenome"}

\item{pileupSubset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}
}
\description{
Checks for max or minimum coverage outside match region
}
\keyword{internal}
