% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifSumm.R
\name{classifSumm}
\alias{classifSumm}
\title{Summarizes best shape-matches into a table}
\usage{
classifSumm(pileup, bestMatchList, windowSize, mode)
}
\arguments{
\item{bestMatchList}{classifications made with shape_matcher function. classifications are stored as the first item in the bestMatchList.}

\item{windowSize}{The number of base pairs to average coverage values over. For compatibility with shape-matching, window sizes can only be 200, 500, 1000 and 2000. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. 1000bp windowSize is the default.}

\item{mode}{Either "genome" or "metagenome"}

\item{pileupSubset}{A table containing pileupSubset names, coverages averaged over 100bp windows, and pileupSubset positions associated with mapping whole-community reads to whole-community pileupSubsets}
}
\description{
Summarizes the classifications made in the shape_matcher function based on which shape match achieved the lowest mean absolute difference for each pileupSubset. Outputs results in a table rather than a list.
}
\keyword{internal}
