% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBestMatchInfo.R
\name{collectBestMatchInfo}
\alias{collectBestMatchInfo}
\title{Collect information regarding the pattern-match}
\usage{
collectBestMatchInfo(pattern, pileupSubset, elevOrGap, leftRightFull)
}
\description{
Make a list containing the match-score, min and max pattern-match values, the start and stop positions of the
elevated or gapped region, the elevation ratio and the classification
}
\keyword{internal}
