% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeprophage_full_func.R
\name{full_prophage_func_WC}
\alias{full_prophage_func_WC}
\title{Full prophage shape-matcher}
\usage{
full_prophage_func_WC(microbial_subset, windowsize)
}
\arguments{
\item{microbial_subset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Build the full prophage shape and provide it as input to the proshape_translator_func to be translated across a contig. The shape is made smaller length-wise 2000bp at a time, and the shape stops decreasing in length once it reaches 10000bp.
}
