% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullElevGapFunctions.R
\name{fullElevGap}
\alias{fullElevGap}
\title{Controller function for full elevation/gap pattern-matching}
\usage{
fullElevGap(pileupSubset, windowSize, minSize, maxSize, elevOrGap)
}
\arguments{
\item{pileupSubset}{A subset of the pileup that pertains only to the contig/chunk
currently being assessed.}

\item{windowSize}{The number of basepairs to average read coverage values over.
Options are 100, 200, 500, 1000 ONLY. Default is 1000.}

\item{minSize}{The minimum size (in bp) of elevation or gap patterns. Default is 10000.}

\item{maxSize}{The maximum size (in bp) of elevation or gap patterns. Default is NA
(i.e. no maximum).}

\item{elevOrGap}{Pattern-matching on 'elevation' or 'gap' pattern.}
}
\description{
Builds full elevation/gap pattern-matches, shrinks the width, and collects
best match information
}
\keyword{internal}
