% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeChunker_func.R
\name{GenomeChunks}
\alias{GenomeChunks}
\title{Break genome file into chunks for pattern-matching analysis}
\usage{
GenomeChunks(genome_pileup, chunk_size)
}
\arguments{
\item{genome_pileup}{Pileup file for read coverages mapped to a genome}

\item{chunk_size}{Value to chunk genome into. Should be divisible by 100. Default is 100,000bp per chunk.}
}
\description{
Break genome file into chunks for pattern-matching analysis
}
\keyword{internal}
