% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkingFunctions.R
\name{genomeChunks}
\alias{genomeChunks}
\title{'chunk' genomes}
\usage{
genomeChunks(pileup, chunkSize)
}
\arguments{
\item{pileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{chunkSize}{If `mode`="genome" OR if `mode`="metagenome" and `chunkContigs`=TRUE,
chunk the genome or contigs, respectively, into smaller subsets for pattern-matching.
`chunkSize` determines the size (in bp) of each 'chunk'. Default is 50000.}
}
\description{
Subset genome pileup into chunks for pattern-matching
}
\keyword{internal}
