% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkingFunctions.R
\name{linkChunks}
\alias{linkChunks}
\title{Link pattern-matches on contig/genome chunks}
\usage{
linkChunks(bestMatchList, pileup, windowSize, mode, verbose)
}
\arguments{
\item{bestMatchList}{A list containing pattern-match information associated with
all contigs/chunks classified by `ProActive()` pattern-matching}

\item{pileup}{A .txt file containing mapped sequencing read coverages averaged over
100 bp windows/bins.}

\item{windowSize}{The number of basepairs to average read coverage values over.}

\item{mode}{Either "genome" or "metagenome"}

\item{verbose}{TRUE or FALSE. Print progress messages to console. Default is TRUE.}
}
\description{
Detect partial gap/elevation pattern matches that fall on the edges of chunked
genomes/contigs that may be part of the same pattern prior to chunking
}
\keyword{internal}
