% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxCoverageCheck_func.R
\name{max_coverage_check}
\alias{max_coverage_check}
\title{Checks for max coverage outside match region}
\usage{
max_coverage_check(best_match_list, microbialread_dataset, windowsize)
}
\arguments{
\item{microbial_subset}{}
}
\description{
Checks for max coverage outside match region
}
\keyword{internal}
