% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noPattern.R
\name{noPattern}
\alias{noPattern}
\title{No transduction pattern-matcher for whole-community read coverage}
\usage{
noPattern(pileupSubset)
}
\arguments{
\item{pileupSubset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}
}
\description{
Assess whether a contig does not have a read coverage pattern associated with a transduction event. A horizontal line at the mean coverage should be an optimal match if the contig read coverage displays no transduction patterns
}
\keyword{internal}
