% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFsInElevGaps.R
\name{ORFsInElevGaps}
\alias{ORFsInElevGaps}
\title{shape-match size calculator}
\usage{
ORFsInElevGaps(elevGapSummList, windowSize, ORFTable)
}
\arguments{
\item{elevGapSummList}{A list containing shape information associated with all contigs containing a potential active prophages. Generated with the allprophages_func}

\item{windowSize}{The window size used to re-average read coverage datasets}

\item{ORFTable}{A gff file of ORFs}
}
\description{
Calculate the size, in base pairs, of the matching region for shapes detected in whole-community read coverages
}
\keyword{internal}
