% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialElevGapFunctions.R
\name{partialElevGap}
\alias{partialElevGap}
\title{Controller function for partial elevation/gap pattern-matching}
\usage{
partialElevGap(pileupSubset, windowSize, minSize, maxSize)
}
\arguments{
\item{pileupSubset}{A subset of the pileup that pertains only to the contig/chunk
currently being assessed.}

\item{windowSize}{The number of basepairs to average read coverage values over.
Options are 100, 200, 500, 1000 ONLY. Default is 1000.}

\item{minSize}{The minimum size (in bp) of elevation or gap patterns. Default is 10000.}

\item{maxSize}{The maximum size (in bp) of elevation or gap patterns. Default is NA
(i.e. no maximum).}
}
\description{
Builds partial elevation/gap pattern-match for patterns going off both the
left and right sides of the contig/chunk, shrinks the width, and collects
best match information
}
\keyword{internal}
