% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialElevGapShrink.R
\name{partialElevGapShrink}
\alias{partialElevGapShrink}
\title{Build elevation pattern, translate it across contig and shrink the width}
\usage{
partialElevGapShrink(
  minCov,
  windowSize,
  maxCov,
  elevLength,
  nonElev,
  bestMatchInfo,
  pileupSubset,
  minSize,
  leftOrRight,
  elevOrGap
)
}
\arguments{
\item{minSize}{The minimum size of elevated read coverage that ProActive searches for. Default is 10000 base pairs.}
}
\description{
Build full-size elevation pattern, run it through the elevation translator function and then shrink the pattern until it reaches the minSize
}
\keyword{internal}
