% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBuilder.R
\name{patternBuilder}
\alias{patternBuilder}
\title{Builds pattern-match vectors}
\usage{
patternBuilder(pileupSubset, bestMatchInfo)
}
\arguments{
\item{pileupSubset}{A subset of the pileup that pertains only to the contig/chunk
currently being assessed.}

\item{bestMatchInfo}{The information associated with the current best pattern-match
for the contig/chunk being assessed.}
}
\description{
Builds the pattern-match (vector) associated with each contig/chunk for
visualization.
}
\keyword{internal}
