% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBuilder.R
\name{patternBuilder}
\alias{patternBuilder}
\title{pattern-builder for patterns detected in whole-community read coverages}
\usage{
patternBuilder(pileupSubset, classifList, i)
}
\arguments{
\item{pileupSubset}{A subset of the read coverage dataset that pertains only to the contig currently being assessed}

\item{classifList}{A list containing pattern information associated with all contigs classified by ProActive}

\item{i}{The index position associated with the current contig's best pattern-match information}
}
\description{
Builds the pattern (vector) associated with the 'best pattern-match' information associated with each contig so it can be plotted.
}
\keyword{internal}
