% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullElevGapFunctions.R
\name{patternTranslator}
\alias{patternTranslator}
\title{Full elevation/gap pattern translator}
\usage{
patternTranslator(contigCov, bestMatchInfo, windowSize, pattern, elevOrGap)
}
\arguments{
\item{contigCov}{The read coverages that pertain to the pileupSubset}

\item{bestMatchInfo}{The information associated with the current best pattern-match
for the contig/chunk being assessed.}

\item{windowSize}{The number of basepairs to average read coverage values over.
Options are 100, 200, 500, 1000 ONLY. Default is 1000.}

\item{pattern}{A vector containing the values associated with the pattern-match}

\item{elevOrGap}{Pattern-matching on 'elevation' or 'gap' pattern.}
}
\description{
Translates full elevation/gap patterns across contigs/chunks 1000bp at a time.
Translation stops when the elevation pattern is 5000bp from the end of the contig/chunk.
}
\keyword{internal}
