% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProActive-package.R
\docType{package}
\name{ProActive-package}
\alias{ProActive-package}
\title{\bold{ProActive}}
\description{
Automatic detection of gaps and elevations in mapped sequencing reads using read coverage
  pattern-matching.
}
\details{
The two main functions in ProActive are:
\enumerate{
\item \code{\link{ProActive}} performs the pattern-matching and characterization read coverage patterns.
\item \code{\link{plotProActiveResults}} plots the results from
 \code{ProActive()}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jlmaier12/ProActive}
  \item Report bugs at \url{https://github.com/jlmaier12/ProActive/issues}
}

}
\author{
Jessie Maier \email{jlmaier@ncsu.edu}
}
\keyword{internal}
