% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProActive.R
\name{ProActive}
\alias{ProActive}
\title{Highly active/abundant MGE prediction function}
\usage{
ProActive(
  pileup,
  mode,
  windowsize = 1000,
  minsize = 10000,
  maxsize = Inf,
  chunksize = 1e+05,
  cleanup = TRUE
)
}
\arguments{
\item{pileup}{A pileup file for either a genome or metagenome containing mapped read coverages averaged over 100 bp window sizes}

\item{mode}{Either "genome" or "metagenome"}

\item{windowsize}{The number of base pairs to average coverage values over. For compatibility with shape-matching, window sizes can only be 200, 500, 1000 and 2000. Larger window sizes improve processing time but the resolution of read coverage patterns may be lost. 1000bp windowsize is the default.}

\item{minsize}{The minimum size of elevated read coverage that ProActive searches for. Default is 10000 base pairs.}

\item{maxsize}{The minimum size of elevated read coverage that ProActive searches for. Default is the close to the length of the contig being assessed.}

\item{chunksize}{If mode="genome", ProActive will split your genome into chunks to search for elevated read coverage. Since ProActive can only identify one elevated read coverage pattern per 'chunk', choosing smaller chunks will increase sensitivity AND processing time. Default is 100,000bp per chunk, with the exception of the last chunk which is the remainder number of base pairs once the genome has been evenly 'chunked'.}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into ProActive with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
The main function that ties together all the other functions. This function performs all the shape-matching and summarizes the results into a list. The first item in the list is a table consisting of the summary information of all the contigs that passed through shape-matching (i.e were not filtered out). The second item in the list is a table consisting of the summary information of all contigs that were predicted as containing a potential active prophage. The third item in the list contains the best shape-match information associated with each contig in the previous table. The fourth and final object in the list is a table containing the contigs that were filtered out prior to shape_matching and the reason why.
}
\examples{
\dontrun{
ProActive_results <- ProActive(pileup=whole_commreadcoverages, mode="metagenome", windowsize=1000, minsize=10000, maxsize=50000, cleanup=TRUE)
}
}
