% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProActivePredictionPlots.R
\name{ProActivePredictionPlots}
\alias{ProActivePredictionPlots}
\title{Plot read coverage graphs of contigs with predicted active prophages in the whole-community read coverages}
\usage{
ProActivePredictionPlots(
  pileup,
  ProActiveResults,
  ProActive_shapelist,
  elevation_filter = NA,
  cleanup = TRUE
)
}
\arguments{
\item{pileup}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{ProActiveResults}{The result from ProActive()}

\item{ProActive_shapelist}{A list containing shape information associated with all contigs containing a potential active prophage. The second, third or fourth item in the list output from ProActive()}

\item{elevation_filter}{Default is NA. The user may define an elevation factor in which any plots with elevated read coverages below the elevation factor are not displayed}

\item{cleanup}{If pileup files are generated by BBMap's pileup.sh, the files output from pileup.sh can be input directly into ProActive with cleanup=TRUE. If your pileup files are NOT generated with BBMap's pileup.sh, set cleanup=FALSE. The user is responsible for data cleaning/formatting if this option is used.}
}
\description{
Plot the read coverages of a contig and its associated shape-match for each contig with a predicted active prophage.
}
\examples{
\dontrun{
ProActivePredictionPlots(whole_commreadcovs, ProActiveResults, ProActiveResults$ExtraConfidentPredictions,cleanup=TRUE)
}
}
