% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCoverageGraphs_activeprophage_func.R
\name{ProActivePredictionPlots}
\alias{ProActivePredictionPlots}
\title{Plot read coverage graphs of contigs with predicted active prophages in the whole-community read coverages}
\usage{
ProActivePredictionPlots(
  microbialread_dataset,
  prophagepredictions,
  final_summary_table,
  windowsize = 1000
)
}
\arguments{
\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{prophagepredictions}{A list containing shape information associated with all contigs containing a potential active prophage. The third item in the list output from the main prophage_activity_finder_func}

\item{final_summary_table}{Prediction summary table with active prophage information. The first item in the list output from the main prophage_activity_finder_func function.}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Plot the read coverages of a contig and its associated shape-match for each contig with a predicted active prophage.
}
\examples{
\dontrun{
ProActivePredictionPlots(whole_commreadcovs, shapematch_result[[3]],shapematch_result[[1]],1000)
}
}
