% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_matching_func_activeprophage.R
\name{shape_matcher_WC}
\alias{shape_matcher_WC}
\title{Shape-matching function for active prophages in whole-community read coverage}
\usage{
shape_matcher_WC(microbialread_dataset, windowsize)
}
\arguments{
\item{microbialread_dataset}{A table containing contig names, coverages averaged over 100bp windows, and contig positions associated with mapping whole-community reads to whole-community contigs}

\item{windowsize}{The window size used to re-average read coverage datasets}
}
\description{
Creates the microbial_subset, representative of one contig, that is used as input for each individual shape-building function. After the information associated with the best match for each shape is obtained, the shape with the lowest mean absolute difference (score) is chosen as the prediction for the contig being assessed.
}
\keyword{internal}
